requirements.txt
flask==3.0.0
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
anthropic==0.7.8
reportlab==4.0.7
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
Pillow==10.1.0
gunicorn==21.2.0
redis==5.0.1
